Updated ID -> UUID migrations to be more sane.
Now there is only one migration file, no SQLite-specific migration file, and
overall a lot less repeat logic. There is still a VERY weird issue going on
as the migration works when testing in MySQL and SQLite individually, but when
both dialets are uncommented in the test_migrations.conf file, the MySQL migration
fails.

Change-Id: I22ba69550ccbc04a24d680748d60414d2d0f1a99